src/pdf_to_markdown_original/pyproject.toml
[project]
name = "pdf-to-markdown-original"
version = "1.1.0"
description = "PDF to Markdown converter (original implementation)"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "PyPDF2>=3.0.0",
    "pdf2image>=1.16.3",
    "Pillow>=10.0.0",
    "markdown==3.5.2",
]
readme = "README.md"
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []
python-version = "3.8"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3 